refactor: remove DQN implementation and replace with DNN model; update example scripts for consistency
