+++ examples/eval_model.py
@@ -1,6 +1,7 @@
 import argparse
 
 import torch
+import torch.nn as nn
 import gymnasium as gym
 
 from rl_tetris.wrapper.Grouped import GroupedStepWrapper
@@ -24,6 +25,31 @@ def get_args():
     return args
 
 
+# DNN 모델
+class DNN(nn.Module):
+    def __init__(self, input_dim=4, hidden_dim=64, output_dim=1):
+        """게임 상태에 대한 휴리스틱한 정보(지워진 줄, 구멍, 인접열 차이 합, 높이 합)를 입력으로 받아서, q-value를 출력하는 DNN 모델"""
+        super(DNN, self).__init__()
+
+        self.model = nn.Sequential(
+            nn.Linear(input_dim, hidden_dim),
+            nn.ReLU(),
+            nn.Linear(hidden_dim, hidden_dim),
+            nn.ReLU(),
+            nn.Linear(hidden_dim, output_dim)
+        )
+
+        self.model.apply(self._init_weights)
+
+    def _init_weights(self, m):
+        if isinstance(m, nn.Linear):
+            nn.init.xavier_uniform_(m.weight)
+            nn.init.constant_(m.bias, 0)
+
+    def forward(self, x):
+        return self.model(x)
+
+
 def test(opt):
     model_path = f"{opt.model_dir}/{opt.model_name}"
 
@@ -32,7 +58,10 @@ def test(opt):
     else:
         torch.manual_seed(42)
 
-    model = torch.load(model_path, map_location=torch.device("cpu")).eval()
+    # 모델 불러오기
+    model = DNN()
+    model.load_state_dict(torch.load(model_path))
+    model.eval()
 
     env = gym.make("RL-Tetris-v0", render_mode="animate")
     env = GroupedStepWrapper(
